moved to store.go
